.archive/spark/custom_image/requirements.txt
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
redis==4.6.0
clickhouse-driver==0.2.6
psycopg2-binary==2.9.7
pyarrow==12.0.1
